# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cipla Healthcare Campaign Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  message: string;
}

export interface LoginInput {
  employeeCode: string;
  password: string;
}

export type AuthUserRole = (typeof AuthUserRole)[keyof typeof AuthUserRole];

export const AuthUserRole = {
  admin: "admin",
  manager: "manager",
} as const;

export interface AuthUser {
  id: number;
  name: string;
  employeeCode: string;
  role: AuthUserRole;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  manager: "manager",
} as const;

export interface User {
  id: number;
  name: string;
  employeeCode: string;
  role: UserRole;
  createdAt: string;
}

export interface ManagerInput {
  name: string;
  employeeCode: string;
  password: string;
}

export interface ManagerWithStats {
  id: number;
  name: string;
  employeeCode: string;
  totalDoctors: number;
  photosUploaded: number;
  documentsUploaded: number;
  completedProfiles: number;
  completionPercentage: number;
  createdAt: string;
}

export interface Doctor {
  id: number;
  managerId: number;
  /** @nullable */
  managerName?: string | null;
  doctorName: string;
  specialization: string;
  city: string;
  /** @nullable */
  clinicAddress?: string | null;
  /** @nullable */
  phoneNumber?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  documentUrl?: string | null;
  /** @nullable */
  photoUploadedAt?: string | null;
  /** @nullable */
  documentUploadedAt?: string | null;
  isComplete: boolean;
  createdAt: string;
}

export interface DoctorInput {
  doctorName: string;
  specialization: string;
  city: string;
  /** @nullable */
  clinicAddress?: string | null;
  /** @nullable */
  phoneNumber?: string | null;
}

export interface DoctorListResponse {
  doctors: Doctor[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface AdminDashboardStats {
  totalDoctors: number;
  totalPhotos: number;
  totalDocuments: number;
  pendingProfiles: number;
  completionPercentage: number;
  totalManagers: number;
  recentAdditions: number;
}

export interface ManagerDashboardStats {
  totalDoctors: number;
  photosUploaded: number;
  documentsUploaded: number;
  completedProfiles: number;
  completionPercentage: number;
}

export interface DailyCount {
  date: string;
  count: number;
}

export interface ManagerPerformance {
  id: number;
  name: string;
  totalDoctors: number;
  completedProfiles: number;
  completionPercentage: number;
}

export type GetAdminDashboardParams = {
  /**
   * @nullable
   */
  startDate?: string | null;
  /**
   * @nullable
   */
  endDate?: string | null;
};

export type ListAllDoctorsParams = {
  /**
   * @nullable
   */
  search?: string | null;
  /**
   * @nullable
   */
  managerId?: number | null;
  /**
   * @nullable
   */
  status?: string | null;
  /**
   * @nullable
   */
  city?: string | null;
  page?: number;
  limit?: number;
  /**
   * @nullable
   */
  startDate?: string | null;
  /**
   * @nullable
   */
  endDate?: string | null;
};

export type GetDailyAdditionsParams = {
  days?: number;
};

export type ListMyDoctorsParams = {
  /**
   * @nullable
   */
  search?: string | null;
  /**
   * @nullable
   */
  status?: string | null;
  /**
   * @nullable
   */
  city?: string | null;
  page?: number;
  limit?: number;
};