# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Cipla Healthcare Campaign Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with employee code and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/dashboard:
    get:
      operationId: getAdminDashboard
      tags: [admin]
      summary: Get admin dashboard statistics
      parameters:
        - name: startDate
          in: query
          schema:
            type: ["string", "null"]
        - name: endDate
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDashboardStats"
  /admin/managers:
    get:
      operationId: listManagers
      tags: [admin]
      summary: List all managers with their stats
      responses:
        "200":
          description: List of managers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ManagerWithStats"
    post:
      operationId: createManager
      tags: [admin]
      summary: Create a new manager
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManagerInput"
      responses:
        "201":
          description: Manager created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/managers/{id}:
    delete:
      operationId: deleteManager
      tags: [admin]
      summary: Delete a manager
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Manager deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/doctors:
    get:
      operationId: listAllDoctors
      tags: [admin]
      summary: List all doctors across all managers
      parameters:
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: managerId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: city
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: startDate
          in: query
          schema:
            type: ["string", "null"]
        - name: endDate
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Paginated list of doctors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorListResponse"
  /admin/daily-additions:
    get:
      operationId: getDailyAdditions
      tags: [admin]
      summary: Get daily doctor addition counts for chart
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Daily additions data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyCount"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    LoginInput:
      type: object
      required: [employeeCode, password]
      properties:
        employeeCode:
          type: string
        password:
          type: string
    AuthUser:
      type: object
      required: [id, name, employeeCode, role]
      properties:
        id:
          type: integer
        name:
          type: string
        employeeCode:
          type: string
        role:
          type: string
          enum: [admin, manager]
    User:
      type: object
      required: [id, name, employeeCode, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        employeeCode:
          type: string
        role:
          type: string
          enum: [admin, manager]
        createdAt:
          type: string
    ManagerInput:
      type: object
      required: [name, employeeCode, password]
      properties:
        name:
          type: string
        employeeCode:
          type: string
        password:
          type: string
    ManagerWithStats:
      type: object
      required: [id, name, employeeCode, totalDoctors, photosUploaded, documentsUploaded, completedProfiles, completionPercentage, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        employeeCode:
          type: string
        totalDoctors:
          type: integer
        photosUploaded:
          type: integer
        documentsUploaded:
          type: integer
        completedProfiles:
          type: integer
        completionPercentage:
          type: number
        createdAt:
          type: string
    Doctor:
      type: object
      required: [id, managerId, doctorName, specialization, city, createdAt, isComplete]
      properties:
        id:
          type: integer
        managerId:
          type: integer
        managerName:
          type: ["string", "null"]
        doctorName:
          type: string
        specialization:
          type: string
        city:
          type: string
        clinicAddress:
          type: ["string", "null"]
        phoneNumber:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        documentUrl:
          type: ["string", "null"]
        photoUploadedAt:
          type: ["string", "null"]
        documentUploadedAt:
          type: ["string", "null"]
        isComplete:
          type: boolean
        createdAt:
          type: string
...
    DoctorListResponse:
      type: object
      required: [doctors, total, page, limit, totalPages]
      properties:
        doctors:
          type: array
          items:
            $ref: "#/components/schemas/Doctor"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
    AdminDashboardStats:
      type: object
      required: [totalDoctors, totalPhotos, totalDocuments, pendingProfiles, completionPercentage, totalManagers, recentAdditions]
      properties:
        totalDoctors:
          type: integer
        totalPhotos:
          type: integer
        totalDocuments:
          type: integer
        pendingProfiles:
          type: integer
        completionPercentage:
          type: number
        totalManagers:
          type: integer
        recentAdditions:
          type: integer
...
    DailyCount:
      type: object
      required: [date, count]
      properties:
        date:
          type: string
        count:
          type: integer
...