# lib/api-zod/src/generated/types/doctor.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cipla Healthcare Campaign Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Doctor {
  id: number;
  managerId: number;
  /** @nullable */
  managerName?: string | null;
  doctorName: string;
  specialization: string;
  city: string;
  /** @nullable */
  clinicAddress?: string | null;
  /** @nullable */
  phoneNumber?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  documentUrl?: string | null;
  /** @nullable */
  photoUploadedAt?: string | null;
  /** @nullable */
  documentUploadedAt?: string | null;
  isComplete: boolean;
  createdAt: string;
}