# lib/api-zod/src/generated/types/doctorListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cipla Healthcare Campaign Management API
 * OpenAPI spec version: 0.1.0
 */
import type { Doctor } from "./doctor";

export interface DoctorListResponse {
  doctors: Doctor[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}