# lib/api-zod/src/generated/types/managerWithStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cipla Healthcare Campaign Management API
 * OpenAPI spec version: 0.1.0
 */

export interface ManagerWithStats {
  id: number;
  name: string;
  employeeCode: string;
  totalDoctors: number;
  photosUploaded: number;
  documentsUploaded: number;
  completedProfiles: number;
  completionPercentage: number;
  createdAt: string;
}